move converter to new file
